refactor: use adapter name
