+++ internal/entitlement/metered/utils_test.go
@@ -9,16 +9,16 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/openmeterio/openmeter/internal/credit"
+	credit_postgres_adapter "github.com/openmeterio/openmeter/internal/credit/adapter"
 	"github.com/openmeterio/openmeter/internal/credit/balance"
 	"github.com/openmeterio/openmeter/internal/credit/grant"
-	credit_postgres_adapter "github.com/openmeterio/openmeter/internal/credit/postgresadapter"
 	"github.com/openmeterio/openmeter/internal/ent/db"
 	"github.com/openmeterio/openmeter/internal/entitlement"
+	entitlement_postgresadapter "github.com/openmeterio/openmeter/internal/entitlement/adapter"
 	meteredentitlement "github.com/openmeterio/openmeter/internal/entitlement/metered"
-	entitlement_postgresadapter "github.com/openmeterio/openmeter/internal/entitlement/postgresadapter"
 	"github.com/openmeterio/openmeter/internal/event/publisher"
 	"github.com/openmeterio/openmeter/internal/productcatalog"
-	productcatalog_postgresadapter "github.com/openmeterio/openmeter/internal/productcatalog/postgresadapter"
+	productcatalog_postgresadapter "github.com/openmeterio/openmeter/internal/productcatalog/adapter"
 	streaming_testutils "github.com/openmeterio/openmeter/internal/streaming/testutils"
 	"github.com/openmeterio/openmeter/internal/testutils"
 	"github.com/openmeterio/openmeter/openmeter/meter"
