+++ internal/productcatalog/adapter/transaction.go
@@ -1,4 +1,4 @@
-package postgresadapter
+package adapter
 
 import (
 	"context"
