+++ test/entitlement/regression/framework_test.go
@@ -7,19 +7,19 @@ import (
 	"time"
 
 	"github.com/openmeterio/openmeter/internal/credit"
+	grantrepo "github.com/openmeterio/openmeter/internal/credit/adapter"
 	"github.com/openmeterio/openmeter/internal/credit/balance"
 	"github.com/openmeterio/openmeter/internal/credit/grant"
-	grantrepo "github.com/openmeterio/openmeter/internal/credit/postgresadapter"
 	"github.com/openmeterio/openmeter/internal/ent/db"
 	"github.com/openmeterio/openmeter/internal/entitlement"
+	entitlementrepo "github.com/openmeterio/openmeter/internal/entitlement/adapter"
 	booleanentitlement "github.com/openmeterio/openmeter/internal/entitlement/boolean"
 	meteredentitlement "github.com/openmeterio/openmeter/internal/entitlement/metered"
-	entitlementrepo "github.com/openmeterio/openmeter/internal/entitlement/postgresadapter"
 	staticentitlement "github.com/openmeterio/openmeter/internal/entitlement/static"
 	"github.com/openmeterio/openmeter/internal/event/publisher"
 	"github.com/openmeterio/openmeter/internal/meter"
 	"github.com/openmeterio/openmeter/internal/productcatalog"
-	productcatalogrepo "github.com/openmeterio/openmeter/internal/productcatalog/postgresadapter"
+	productcatalogrepo "github.com/openmeterio/openmeter/internal/productcatalog/adapter"
 	streamingtestutils "github.com/openmeterio/openmeter/internal/streaming/testutils"
 	"github.com/openmeterio/openmeter/internal/testutils"
 	"github.com/openmeterio/openmeter/pkg/models"
